Move color helper to utilities
