+++ frontend/src/utilities/helpers.js
@@ -132,3 +132,16 @@ export const removeHtmlTags = (str) => {
   if (!str || !str.toString) return str;
   return str.toString().replace(/<\/?[a-z]+>/gi, '');
 };
+
+export function hexToRGB(hex, alpha) {
+  if (!hex) return 'rgba(0, 0, 0, 0)';
+  const r = parseInt(hex.slice(1, 3), 16),
+    g = parseInt(hex.slice(3, 5), 16),
+    b = parseInt(hex.slice(5, 7), 16);
+
+  if (alpha) {
+    return 'rgba(' + r + ', ' + g + ', ' + b + ', ' + alpha + ')';
+  } else {
+    return 'rgb(' + r + ', ' + g + ', ' + b + ')';
+  }
+}
