+++ frontend/src/views/components/Journey/Chapter/component.tsx
@@ -1,5 +1,7 @@
 import type { JourneyStep as StaticJourneyStep } from 'types/static-journeys';
 import type { JourneyAttributes } from 'types/journeys';
+import { hexToRGB } from 'utilities/helpers';
+
 const STATIC_JOURNEYS = process.env.NEXT_PUBLIC_STATIC_JOURNEYS === 'true';
 
 const StaticChapter: React.FC<StaticJourneyStep> = ({ background, title, content }) => (
@@ -14,19 +16,6 @@ const StaticChapter: React.FC<StaticJourneyStep> = ({ background, title, content
     </div>
   </div>
 );
-
-function hexToRGB(hex: string, alpha: number) {
-  if (!hex) return 'rgba(0, 0, 0, 0)';
-  const r = parseInt(hex.slice(1, 3), 16),
-    g = parseInt(hex.slice(3, 5), 16),
-    b = parseInt(hex.slice(5, 7), 16);
-
-  if (alpha) {
-    return 'rgba(' + r + ', ' + g + ', ' + b + ', ' + alpha + ')';
-  } else {
-    return 'rgb(' + r + ', ' + g + ', ' + b + ')';
-  }
-}
 
 const Chapter: React.FC<JourneyAttributes> = ({
   background_image: backgroundImage,
